% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_data_gen.R
\name{tensor_data_gen}
\alias{tensor_data_gen}
\title{Tensor time series data generation.}
\usage{
tensor_data_gen(K, T, d, r, re, eta, u)
}
\arguments{
\item{K}{The number of modes for the tensor time series.}

\item{T}{Length of time series.}

\item{d}{Dimensions of each mode of the tensor, written in a vector of length \code{K}.}

\item{r}{Rank of core tensors, written in a vector of length \code{K}.}

\item{re}{Dimension of the cross-sectional common error.}

\item{eta}{Quantities controlling factor strengths in each factor loading matrix, written in a list of \code{K} vectors.}

\item{u}{Quantities controlling range of elements in each factor loading matrix, written in a list of \code{K} vectors.}
}
\value{
A list containing the following: \cr
\code{X}: the generated tensor time series, stored in a 'Tensor' object defined in \pkg{rTensor}, where mode-1 is the time mode \cr
\code{A}: a list of K factor loading matrices \cr
\code{F}: time series of core tensor, stored in a 'Tensor' object, where mode-1 is the time mode \cr
\code{E}: time series of error tensor, stored in a 'Tensor' object, where mode-1 is the time mode \cr
}
\description{
Function to generate a random sample of time series tensor factor model, based on econometrics assumptions.
}
\details{
Input tensor dimension and rank of core tensor, return a sample of tensor time series generated by factor model.
}
\examples{
\dontrun{
set.seed(10)
K = 2
T = 100
d = c(40,40)
r = c(2,2)
re = 10
eta = list(c(0,0),c(0,0))
u = list(c(-2,2),c(-2,2))
Data_test = tensor_data_gen(K,T,d,r,re,eta,u)
print(Data_test$X)
}
}
